pyproject.toml
[build-system]
requires = [ "setuptools>=61.0" ]
build-backend = "setuptools.build_meta"


[project]
name = "lmdbsystem"
version = "0.1.4"
description = "Abstraction for read and write to the lmdb file"
readme = "README.md"
keywords = [ "lmdbsystem", "lmdb" ]
authors = [
    { name = "Dao Quang Duy", email = "duydaoquang12@gmail.com" },
]
maintainers = [
    { name = "Rabiloo Developers", email = "oss@rabiloo.com" },
]
license = {text = "The MIT License"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 2 - Pre-Alpha",
]
dependencies = [
    "lmdb",
    "Pillow",
    "opencv-python-headless",
    "numpy",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "flake8-pyproject",
    "pytest",
    "pytest-asyncio",
    "typing_extensions",
    "isort"
]

[project.urls]
"Homepage" = "https://github.com/rabiloo/python-lmdbsystem"
"Repository" = "https://github.com/rabiloo/python-lmdbsystem"
"Bug Tracker" = "https://github.com/rabiloo/python-lmdbsystem/issues"


[tool.flake8]
count = true
max-complexity = 15
max-line-length = 120
statistics = true
ignore = ["W503"]


[tool.black]
target-version = ["py310"]
line-length = 120


[tool.isort]
profile = "black"
lines_between_types = 1
line_length = 120
